Bug fixes.
Fixed some bugs.
Line 95 in device.js - wrong parameter list to the function passed as third parameter to processMessage().
Also line 82 was double stringifying the JSON.  I think that's it...  Tough to diff because I was making changes to the code in the lib directory in the debugger, which was babled, but checking changes into src directory.  Sorry...